Fix bug in initial iOS isConnected fetch
In iOS, the onConnectivityChange callback in the fetch in componentDidMount (which seems to always report `isConnected = false`) fires AFTER the event listener gets called, causing the device monitor to report old data and false negatives.

This update makes sure that onConnectivityChange only gets called from the componentWillMount callback if it has not received a `change` event from the event listener.